made mistake on file content and fixed and added correct code
